nfc-health-system/requirements.txt
flask
pyjwt
werkzeug
python-dotenv
requests
supabase
